chore(test): Integration Codecov coverage reporting for notebook-controller
This commit integrates Codecov into our testing process by publishing our coverage report generated during `make test` to Codecov via the official `codecov/codecov-action` GHA.

The `CODECOV_TOKEN` was previously uploaded to our GH repository as a secret (and copied to the Dependabot Secrets as well) by a repo admin:
- https://redhat-internal.slack.com/archives/C060A5FJEAD/p1733334721701649

Please be aware of functionality **NOT** included in this initial PR (that may be added in the future):
- unit test results are not being shipped to `Codecov` (even though that capability is supported)
- a `.codecov.yml` file is presently not included.  this could come as a fast-follow after the team gets a "feel" for `Codecov` - but I think its a bit premature to take my personal opinions to create such a file.  Example `odh-dashboard` config can be seen here:
    - https://github.com/opendatahub-io/odh-dashboard/blob/main/.codecov.yml

The `Codecov` integration was added to the `notebook_controller_unit_test.yaml` GHA - which is already configured to execute on `pull_request` and `push` triggers.  As such, no additional modification was necessary.

:warning: It should be noted that the `make test` target as configured in upstream is presently incorrect.  The 2 different `go test` executions are redundant - and the 2nd invocation (targetting only the `./controller/...` scope, actually overwrites the `cover.out` file initially generated by the 1st invocation.  However, the fix for this should come from upstream.. I will open up a PR against upstream at which point it will flow through to `odh` whenever its merged and we sync.

Related-to: https://issues.redhat.com/browse/RHOAIENG-11142