[RHOAIENG-33609] fix route search to work also for generated routes
In the `createOAuthClient` function, the search for the appropriate
route for the workbench didn't work when the generated route was created
before. We're now searching based on the appropriate labels as is used
for other resources in this case when generated name is used instead of
the regular one.

This is a followup of the #539.