[RHOAIENG-24545] Assure proper runtime images configmap presence
Before this patch, the configmap of the runtime images configuration for
the Jupyter IDE wasn't present on a first workbench created in a
namespace. One had to restart the workbench or create another one.
Reason for this was the order of the operations - at the time when
webhook wanted to mount the runtime images configmap, that configmap
hasn't been created yet. This patch assures that it's created in time
for addition even in case of a first workbench start.