Move back the Go version in go.mod file to 1.24
The reason we're doing this is to allow compilation in the internal
build system that doesn't allow us to use Go 1.25 right now. We'll bump
this back to 1.25.z as soon as we can.

For the build itself, we're using the Go toolset 1.25 builder so the
result code is still compiled with the latest binaries avoiding the CVEs
in the standard libraries to be brought into the final binary. So far we
don't use any Go 1.25 features so we're not blocked by this change.